Fix comment in example code in README
The comment claims that the ErrorHandler falls back to the PlainFormatter, when no Formatter matches the request, but that is not true. It falls back to the first formatter in the list. When you do not supply any formatters to the constructor the default list of formatters indeed starts with the PlainFormatter, but that is not what happens in the example.